revisiting binary exponentiation to calculate the pow(x,n) efficiently

+++ Pow(x,n)/main.cpp
@@ -9,12 +9,12 @@ double mypow(double x,int n){
     if (x == -1 && n % 2 == 0) return 1.0;
     if (x == -1 && n % 2 != 0) return -1.0;
 
-    int binForm{n};
+    long binForm{n};
     if(n < 0 ){
         x = 1/x;
         binForm = -binForm;
     }
-    int ans{1};
+    double ans{1};
     while(binForm > 0){
         if(binForm % 2 == 1){
             ans = ans * x;
